Keep Computation FLOPs term consistent
Super extra tiny point: not sure if there was a distinction between Algorithm vs Computation FLOPs. Caused me a moment of confusion, so might be helpful to clarify.

Same idea for `Network Bandwidth`. Below, there's a section titled `Network Bandwidth`, so that makes sense, but here, seems like we're still in the general regime of data transfer bandwidth.